refactor: 내 위치 불러오는 코드 custom hook 분리
